requirements.txt
streamlit
python-dotenv
openai
langchain
langchain-openai
langchain-community
langchain-text-splitters
sentence-transformers
transformers
torch
faiss-cpu
rank-bm25
tiktoken
